mindhub/backend-django/requirements.txt
# Django Core
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-extensions>=3.2.0

# Database
psycopg2-binary>=2.9.5
django-environ>=0.10.0

# Authentication & Security
django-allauth>=0.54.0
django-guardian>=2.4.0
cryptography>=41.0.0

# API & Documentation
drf-spectacular>=0.26.0
django-filter>=23.0.0

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
celery>=5.3.0
redis>=4.5.0

# Development & Production
gunicorn>=21.0.0
whitenoise>=6.5.0
django-debug-toolbar>=4.1.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0

# File Processing
Pillow>=10.0.0
openpyxl>=3.1.0
reportlab>=4.0.0

# Frontend & UI
django-tailwind>=3.6.0
django-crispy-forms>=2.0
crispy-tailwind>=0.5.0

# Monitoring & Logging
sentry-sdk>=1.28.0